Register offenses for constraint creation without `not_valid:true` option
Scanning a large table to verify a new foreign key or check constraint can take a
long time, and other updates to the table are locked out until the
`ALTER TABLE ADD CONSTRAINT` command is committed. Using `NOT VALID` constraint
option is reduce the imact of adding a new constraint on concurrent updates. With
`NOT VALID`, the `ADD CONSTRAINT` command does not scan the table and can be
committed immediately.