CI: Trigger GitHub Actions on fewer events
When a pull request was opened, we would previously run all CI twice -
once for the `push` event when the branch is pushed up, and once for the
`pull_request` event when the PR is opened.

Now we will run CI only once when the PR is opened (or re-opened, or
"synchronized" which I assume is when you push new changes). And we will
run CI for the master branch when a PR is merged.

Some repositories configure CI to run _only_ on the `push` event, but
that doesn't work when making changed on a forked repository.